Move hello_world.tflite to examples/hello_world.
