+++ tensorflow/lite/micro/models/BUILD
@@ -20,14 +20,3 @@ generate_cc_arrays(
     out="person_detect_model_data.h",
 )
 
-generate_cc_arrays(
-    name="generated_hello_world_model_cc",
-    src="hello_world.tflite",
-    out="hello_world_model_data.cc",
-)
-
-generate_cc_arrays(
-    name="generated_hello_world_model_hdr",
-    src="hello_world.tflite",
-    out="hello_world_model_data.h",
-)
